### 2025-11-17 #9450 @editorialbot
[PRE REVIEW]: mosaix-pde: Differentiable pattern-forming PDEs for machine learning, optimization, and control
**Submitting author:** @acoh64 (<a href="http://orcid.org/0000-0002-5284-6775">Alexander Cohen</a>)
**Repository:** https://github.com/acoh64/mosaix-pde
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
