### 2025-11-17 #9450 @kyleniemeyer
Regarding the tests, I see a `tests` directory, but could you add some brief instructions on running these to the README/docs? (Even if just `pytest`.)
