### 2026-01-29 #9450 @fruzsinaagocs
:wave: @ricalessandri, would you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
