### 2026-04-17 #9926 @streeve
Hey, sorry for the delay on my end. Here's a first pass for both the repo and paper.

## Development history
- Clarification: is this still an active project? This seems to be near the line of sufficiently active/open development @fruzsinaagocs can you provide some guidance here?

## Functionality
- Installation: I used a simple `venv` and the installation was as expected. This could be added to the install documentation
- Functionality: the `opt_model.train` from the readme failed (opened https://github.com/acoh64/mosaix-pde/issues/1)

## Documentation
- A statement of need: I would like more clarity on the intended users: any domain scientist looking to solve a problem that relies on PDEs? All the examples so far seem to be in physics/materials
- Example usage: are the `doc/notebooks/` the primary examples? This should be explained in the readme and readthedocs. Ideally examples outside of notebooks should be available. Related, the example in the readme file could also be added in `examples/` and referenced
- Automated testing: minimal, but exists. If runnable examples are added, they should be checked here
- Community guidelines: a CONTRIBUTING.md or similar communication should be added

## Software paper
- Many required sections are missing and must be added
- The statement of need should come earlier, in order to help clarify the background
- Most DOIs are missing
- The performance results should be enlarged and described with a bit more detail (e.g. hardware architectures)
